+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node ./index.js",
-    "start:dev": "nodemon --exec ts-node ./index.js",
+    "start:dev": "nodemon --exec ts-node ./index.ts",
     "start:ws": "nodemon --exec ts-node ./src/ws/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
